+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -88,6 +88,10 @@ class VideoModule : Module() {
         view.videoPlayer?.requiresLinearPlayback = linearPlayback
       }
 
+      Prop("useExoShutter") { view: VideoView, useExoShutter: Boolean? ->
+        view.useExoShutter = useExoShutter ?: true
+      }
+
       AsyncFunction("enterFullscreen") { view: VideoView ->
         view.enterFullscreen()
       }
@@ -113,6 +117,9 @@ class VideoModule : Module() {
       }
 
       OnViewDidUpdateProps { view ->
+        view.useExoShutter = view.useExoShutter ?: true
+        view.applySurfaceViewVisibility()
+
         if (view.playerView.useController != view.useNativeControls) {
           view.playerView.useController = view.useNativeControls
         }